// UI/ChathireUI/src/app/api/model/user-details-dto-for-update.ts
/**
 * REST API
 * Rest Calls
 *
 * The version of the OpenAPI document: v2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './user-dto';


export interface UserDetailsDtoForUpdate { 
    userDtoForUpdate?: UserDto;
    consultancyID?: number | null;
    publicProfileUserName?: string | null;
}